+++ examples/3_auth_callback.rs
@@ -7,7 +7,7 @@
 use env_logger::Builder as EnvLogger;
 use log::{LevelFilter, debug, info, warn};
 use std::borrow::Cow;
-use tss2_fapi_rs::{AuthCbParam, BaseErrorCode, ErrorCode, FapiCallbacks, FapiContext, KeyFlags};
+use tss2_fapi_rs::{BaseErrorCode, Callbacks, ErrorCode, FapiContext, KeyFlags};
 
 const MY_KEYFLAG: &[KeyFlags] = &[KeyFlags::Sign, KeyFlags::NoDA];
 const MY_KEYPATH: &str = "HS/SRK/myTestKey";
@@ -44,7 +44,8 @@ fn main() {
     info!("FAPI context created. ({})", context);
 
     // Add authorization callback function
-    match context.set_callbacks(MyCallbacks { password: MY_AUTHVAL }) {
+    // Will be called by the FAPI in order to request authorization values!
+    match context.set_callbacks(Callbacks::with_auth(|_| Some(Cow::Borrowed(MY_AUTHVAL)))) {
         Ok(_) => info!("Success."),
         Err(error) => panic!("Failed to set up AUTH callback function: {:?}", error),
     }
@@ -81,20 +82,4 @@ fn main() {
 
     // Exit
     info!("Shutting down...");
-}
-
-/// Struct to implement our application-defined callback functions
-#[derive(Debug)]
-struct MyCallbacks {
-    password: &'static str,
-}
-
-impl FapiCallbacks for MyCallbacks {
-    /// This function will be called by FAPI in order to request authorization values from the application.
-    ///
-    /// *Note:* For simplicity, in this example, the callback function always returns our password, regardless of the requested object path.
-    fn auth_cb(&self, param: AuthCbParam) -> Option<Cow<'static, str>> {
-        info!("Authorization for object at {:?} has been requested.", param.object_path);
-        Some(Cow::from(self.password))
-    }
 }